requirements.txt
asgiref==3.3.4
Django==3.2.3
django-colorfield==0.4.1
django-debug-toolbar==3.2.1
django-filter==2.4.0
djangorestframework==3.12.4
flake8==3.9.2
flake8-broken-line==0.3.0
flake8-isort==4.0.0
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
isort==5.8.0
mccabe==0.6.1
pdfkit==0.6.1
pep8-naming==0.11.1
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
sqlparse==0.4.1
testfixtures==6.17.1
